pyproject.toml
[project]
name = "nonebot-plugin-helpwithpic"
version = "1.6"
description = "nonebot插件-多模ai工具"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["PPT", "chat"]
authors = [
  {email = "3173707804@qq.com"},
  {name = "syagina"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

dependencies = [
  "nonebot2",
  "anyio",
  "nonebot-adapter-onebot",
  "aiohttp",
  "pptx",
  "docx",
  "pillow"
]

[project.urls]
repository = "https://github.com/syagina/nonebot-plugin-multigpt"
